% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calcAdjustFactor}
\alias{calcAdjustFactor}
\title{Generate adjust factors based on data after scaling}
\usage{
calcAdjustFactor(
  data,
  spike.in.prefix = NULL,
  enrich.group,
  input.id = "Input",
  enrich.id = "Enrich",
  scale.factor,
  prop.top.enrich = 1,
  top.down = TRUE,
  pseudo.count = 1
)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the
number of samples and p is the number of features.}

\item{spike.in.prefix}{A character specify the prefix of spike-in id, e.g.,
"^FB" stands for gene id of fly spike-in, default: NULL.}

\item{enrich.group}{Vector of enrichment group, e.g., c("Input","Enrich","Input","Enrich").}

\item{input.id}{Input library id, must be consistent with \code{enrich.group}, e.g., "Input".}

\item{enrich.id}{Enrich library id, must be consistent with \code{enrich.group}, e.g., "Enrich".}

\item{scale.factor}{Vector of scale factors generated from \code{calcScaleFactor}.}

\item{prop.top.enrich}{Proportion of top-enriched genes to use for
adjustment of non-specific enrichment, by default all genes are used.}

\item{top.down}{Whether using top-to-down or down-to-top enriched genes for
the adjustment of non-specific enrichment, default: TRUE (top-down)}

\item{pseudo.count}{A numeric scalar of pseudo-counts to be added to each gene, default: 1.}
}
\value{
Vector of adjust factors
}
\description{
Generate adjust factors based on data after scaling
}
