% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Compass-class}
\alias{Compass-class}
\alias{createCompass}
\title{Compass object and constructor}
\usage{
createCompass(
  data,
  col.data,
  spike.in.prefix = NULL,
  input.id = "Input",
  enrich.id = "Enrich",
  synthetic.id = NULL
)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the
number of samples and p is the number of features.}

\item{col.data}{\code{data.frame} with at least two columns (indicate condition and enrich groups).
Rows of \code{col.data} correspond to columns of \code{data}.}

\item{spike.in.prefix}{A character specify the prefix of spike-in id, e.g., "^FB" stands for gene id of fly spike-in , default: NULL.}

\item{input.id}{Input library id, must be consistent with \code{enrich.group}, e.g., "Input".}

\item{enrich.id}{Enrich library id, must be consistent with \code{enrich.group}, e.g., "Enrich".}

\item{synthetic.id}{Vector of synthetic RNA id, e.g. c("Syn1","Syn2"), default: NULL.}
}
\value{
Compass object
}
\description{
\code{Compass} object extends the \code{SummarizedExperiment} class.
The \code{createCompass} is a easy constructor of \code{Compass} object
}
\details{
Description of each slot:
\code{assay} \code{SummarizedExperiment::Assays} object, contains all counts.
\code{counts} list for holding raw/normalized counts of sample and spike_in.
\code{norm_factors} list for holding normalization factors of sample and spike_in.
\code{ratio} list for holding the ratio of all genes.
\code{ratio_filtered} lists for holding the ratio of genes with ratios smaller than 1 in all samples.
\code{parameter} list of parameters.
}
