% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{listParameter}
\alias{listParameter}
\alias{getParameter}
\alias{listParameter,Compass-method}
\alias{getParameter,Compass,character-method}
\title{Accessor of Compass parameter}
\usage{
listParameter(object)

getParameter(object, name)

\S4method{listParameter}{Compass}(object)

\S4method{getParameter}{Compass,character}(object, name)
}
\arguments{
\item{object}{Compass.}

\item{name}{Name of the parameter.}
}
\value{
Vector of parameter names

Vector of parameter
}
\description{
List all parameters.

Get specific parameter.
}
