% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ratioShrinkage}
\alias{ratioShrinkage}
\title{Empirical Bayes Shrinkage of ratio}
\usage{
ratioShrinkage(ratio, bio.group = NULL)
}
\arguments{
\item{ratio}{matrix of ratio, with genes in rows and samples in columns.}

\item{bio.group}{Vector of biology group, e.g., c("Young","Young","Old","Old").}
}
\value{
list of shrunk ratio
}
\description{
Empirical Bayes Shrinkage of ratio
}
