% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{synRatio}
\alias{synRatio}
\title{Calculation of synthetic spike-in ratio}
\usage{
synRatio(object, syn.meta, ratio.shrinkage = TRUE)
}
\arguments{
\item{object}{Compass object.}

\item{syn.meta}{data.frame of metadata of synthetic spike-ins.}

\item{ratio.shrinkage}{Whether to perform shrinkage of ratio, default TRUE.}
}
\value{
data.frame of synthetic spike-in ratio
}
\description{
Calculation of synthetic spike-in ratio
}
