% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CompassAnalyze}
\alias{CompassAnalyze}
\title{Normalization and ratio assessment}
\usage{
CompassAnalyze(
  object,
  adjust = TRUE,
  prop.top.enrich = 1,
  top.down = TRUE,
  pseudo.count = 1
)
}
\arguments{
\item{object}{Compass object}

\item{adjust}{Whether to perform linear regression-based adjustment, default: TRUE}

\item{prop.top.enrich}{Proportion of top-enriched genes to use for
adjustment of non-specific enrichment, by default all genes are used.}

\item{top.down}{Whether using top-to-down or down-to-top enriched genes for
the adjustment of non-specific enrichment, default: TRUE (top-down)}

\item{pseudo.count}{A numeric scalar of pseudo-counts to be added to each gene, default: 1.}
}
\value{
Compass object
}
\description{
Normalization and ratio assessment
}
